Move add* funcitons to python modules
The add* (i.e. addFatras, addPythia8) were moved to modules
in Examples/python/acts/examples.
Two modules were made: simulation hosting generation, simulation and digization
and reconstruciton.

They can be used as follows:
from acts.examples.simulation import addFatras

Example scripts are refactored to use the new location of add* functions.
