+++ Examples/Python/CMakeLists.txt
@@ -56,6 +56,8 @@ target_link_libraries(ActsPythonBindings PUBLIC
 set(py_files 
   __init__.py 
   examples/__init__.py
+  examples/simulation.py
+  examples/reconstruction.py
   _adapter.py
 )
