+++ Examples/Scripts/Python/full_chain_itk.py
@@ -10,11 +10,8 @@ detector, trackingGeometry, decorators = itk.buildITkGeometry(geo_dir)
 field = acts.examples.MagneticFieldMapXyz(str(geo_dir / "bfield/ATLAS-BField-xyz.root"))
 rnd = acts.examples.RandomNumbers(seed=42)
 
-from particle_gun import addParticleGun, MomentumConfig, EtaConfig, ParticleConfig
-from fatras import addFatras
-from digitization import addDigitization
-from seeding import addSeeding, TruthSeedRanges
-from ckf_tracks import addCKFTracks, CKFPerformanceConfig
+from acts.examples.simulation import addParticleGun, MomentumConfig, EtaConfig, ParticleConfig, addFatras, addDigitization
+from acts.examples.reconstruction import addSeeding, TruthSeedRanges, addCKFTracks, CKFPerformanceConfig
 
 s = acts.examples.Sequencer(events=100, numThreads=-1)
 s = addParticleGun(
