+++ Examples/Scripts/Python/full_chain_odd.py
@@ -20,12 +20,8 @@ detector, trackingGeometry, decorators = getOpenDataDetector(mdecorator=oddMater
 field = acts.ConstantBField(acts.Vector3(0.0, 0.0, 2.0 * u.T))
 rnd = acts.examples.RandomNumbers(seed=42)
 
-from particle_gun import addParticleGun, MomentumConfig, EtaConfig, ParticleConfig
-from fatras import addFatras
-from digitization import addDigitization
-from seeding import addSeeding, TruthSeedRanges
-from ckf_tracks import addCKFTracks, CKFPerformanceConfig
-from vertex_fitting import addVertexFitting, VertexFinder
+from acts.examples.simulation import addParticleGun, MomentumConfig, EtaConfig, ParticleConfig, addFatras, addDigitization
+from acts.examples.reconstruction import addSeeding, TruthSeedRanges, addCKFTracks, CKFPerformanceConfig, addVertexFitting, VertexFinder
 
 s = acts.examples.Sequencer(events=100, numThreads=-1, logLevel=acts.logging.INFO)
