+++ Examples/Scripts/Python/truth_tracking_gsf.py
@@ -6,38 +6,7 @@ from acts.examples import Sequencer, GenericDetector, RootParticleReader
 
 import acts
 
-from acts import UnitConstants as u
-
-
-def addTruthTrackingGsf(
-    s: acts.examples.Sequencer,
-    trackingGeometry: acts.TrackingGeometry,
-    field: acts.MagneticFieldProvider,
-):
-    gsfOptions = {
-        "maxComponents": 12,
-        "abortOnError": False,
-        "disableAllMaterialHandling": False,
-    }
-
-    gsfAlg = acts.examples.TrackFittingAlgorithm(
-        level=acts.logging.INFO,
-        inputMeasurements="measurements",
-        inputSourceLinks="sourcelinks",
-        inputProtoTracks="prototracks",
-        inputInitialTrackParameters="estimatedparameters",
-        outputTrajectories="gsf_trajectories",
-        directNavigation=False,
-        pickTrack=-1,
-        trackingGeometry=trackingGeometry,
-        fit=acts.examples.TrackFittingAlgorithm.makeGsfFitterFunction(
-            trackingGeometry, field, **gsfOptions
-        ),
-    )
-
-    s.addAlgorithm(gsfAlg)
-
-    return s
+u = acts.UnitConstants
 
 
 def runTruthTrackingGsf(
@@ -51,10 +20,8 @@ def runTruthTrackingGsf(
     s=None,
 ):
 
-    from particle_gun import addParticleGun, EtaConfig, PhiConfig, ParticleConfig
-    from fatras import addFatras
-    from digitization import addDigitization
-    from seeding import addSeeding, SeedingAlgorithm
+    from acts.examples.simulation import addParticleGun, EtaConfig, PhiConfig, ParticleConfig, addFatras, addDigitization
+    from acts.examples.reconstruction import addSeeding, SeedingAlgorithm, addTruthTrackingGsf
 
     s = s or acts.examples.Sequencer(
         events=100, numThreads=-1, logLevel=acts.logging.INFO
